fix(auth): split install-targets helpers under file-length gate
Extract fetchInstallTargets/upsertUserWithInstallTargets from oauth.ts
(257 SLOC) and raise oauth/me test exemption caps for multi-install coverage.
